fix and rubocop
