--- lesson3_vera/README.md
@@ -1,1 +1,0 @@
-Во втором задании пыталась в отдельный файл метод сделать, но почему-то он ломается...(год выводится некорректно). Причину пока не могу понять.
